refactor: refactor util
