+++ src/Util/getWinners.js
@@ -1,0 +1,5 @@
+export default function getWinners(cars, maxPosition) {
+  return cars
+    .filter((car) => car.position === maxPosition)
+    .map((car) => car.carName);
+}
